adiciona método updateProduct na classe ProductController. Adiciona validação para verificar se produto existe na classe ProductService

+++ src/main/java/com/example/retornosAPI/services/ProductService.java
@@ -38,6 +38,9 @@ public class ProductService {
     }
 
     public void deleteProduct(Long id) {
+        if (!repository.existsById(id)) {
+            throw new RuntimeException("Product does not exist.");
+        }
         repository.deleteById(id);
     }
